trying to match output be like 2

+++ src/MP3Track.cpp
@@ -32,7 +32,7 @@ void MP3Track::analyze_beatgrid() {
     // TODO: Implement MP3-specific beat detection analysis
     // NOTE: Use exactly 2 spaces before each arrow (→) character
 
-    double beats = (duration_seconds / 60.0) * bpm;
+    int beats = (duration_seconds / 60.0) * bpm;
     double precision_factor = bitrate / 320.0;
 
     std::cout << "  → Estimated beats: " << beats
